+++ src-interface/recorder/recorder.cpp
@@ -336,9 +336,9 @@ namespace satdump
 
         if (is_processing)
         {
-            int y_pos = ImGui::GetCursorPosY() + 10 * ui_scale;
-            int live_width = recorder_size.x + 16 * ui_scale;
-            int live_height = 250;
+            float y_pos = ImGui::GetCursorPosY() + 10 * ui_scale;
+            float live_width = recorder_size.x + 16 * ui_scale;
+            float live_height = 250;
             float winwidth = live_pipeline->modules.size() > 0 ? live_width / live_pipeline->modules.size() : live_width;
             float currentPos = 0;
             for (std::shared_ptr<ProcessingModule> module : live_pipeline->modules)
